Déplace le formulaire relatif à l'aide avec sa vue
